超时处理都修正

+++ src/crawling_sub_comment.py
@@ -1,12 +1,15 @@
 import json
 import os
+import threading
 
 import requests
 import analysis
 import connect_mysql
+from src import time_out_exception
 
 
 def crawling_sub_comment_by_id(comment_id, is_print, sub_comment_url):
+    response = None
     param_from = 0
     headers = {
         'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/97.0.4692.71 Safari/537.36 '
@@ -21,22 +24,33 @@ def crawling_sub_comment_by_id(comment_id, is_print, sub_comment_url):
             'from': str(param_from),
             'X-UA': 'V=1&PN=WebAppIntl&LANG=zh_TW&VN_CODE=59&VN=0.1.0&LOC=CN&PLT=PC&DS=Android&UID=22a3964e-db21-41a2-852d-f30283cda0f3&VID=434092598&DT=PC'
         }
-        response = requests.get(url=sub_comment_url, headers=headers, params=params)
-        response_json = response.json()
-        if len(response_json['data']['list']) != 0:
-            response_jsons.append(response_json)
-        if is_print:
-            file_name = './sub_comment_container/sub_comment parent_comment_id = ' + str(comment_id) + ' from = ' + str(
-                param_from) + '.json'
-            with open(file_name, 'w', encoding='utf-8') as fp:
-                json.dump(response_json, fp=fp, ensure_ascii=False, indent=4)
-            print(file_name, "Saved!!!")
+        try:
+            timer = threading.Timer(20, time_out_exception.throw_time_out_error,
+                                    'requests get sub_comment timeout')
+            timer.start()
+            response = requests.get(url=sub_comment_url, headers=headers, params=params)
+            timer.cancel()
+        except TimeoutError as e:
+            print(e)
+        if response is not None:
+            response_json = response.json()
+            response = None
+            if len(response_json['data']['list']) != 0:
+                response_jsons.append(response_json)
+            if is_print:
+                file_name = './sub_comment_container/sub_comment parent_comment_id = ' + str(comment_id) + ' from = ' + str(
+                    param_from) + '.json'
+                with open(file_name, 'w', encoding='utf-8') as fp:
+                    json.dump(response_json, fp=fp, ensure_ascii=False, indent=4)
+                print(file_name, "Saved!!!")
+                if len(response_json['data']['list']) == 0:
+                    os.remove(file_name)
             if len(response_json['data']['list']) == 0:
-                os.remove(file_name)
-        if len(response_json['data']['list']) == 0:
-            return response_jsons
+                return response_jsons
+            else:
+                param_from += 10
         else:
-            param_from += 10
+            param_from+=10
 
 
 def crawling_sub_comment_by_ids_and_app_ids(ids_and_app_ids, is_print, db, sub_comment_url):
